Added users to todos && Changed response format

+++ internal/repository/mongorep/todo.go
@@ -19,15 +19,20 @@ func NewMongoToDoRepository(db *mongo.Database) *ToDoRepository {
 	}
 }
 
-func (rep *ToDoRepository) Get(ctx context.Context, id string) (domain.Todo, error) {
+func (rep *ToDoRepository) Get(ctx context.Context, id, userId string) (domain.Todo, error) {
 	objId, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
+		return domain.Todo{}, err
+	}
+
+	userObjId, err := primitive.ObjectIDFromHex(userId)
 	if err != nil {
 		return domain.Todo{}, err
 	}
 
 	var todo domain.Todo
 	err = rep.db.Collection("todos").
-		FindOne(ctx, bson.M{"_id": objId}).
+		FindOne(ctx, bson.M{"_id": objId, "user_id": userObjId}).
 		Decode(&todo)
 
 	if err != nil {
@@ -37,13 +42,18 @@ func (rep *ToDoRepository) Get(ctx context.Context, id string) (domain.Todo, err
 	return todo, nil
 }
 
-func (rep *ToDoRepository) GetAll(ctx context.Context) ([]domain.Todo, error) {
-	cursor, err := rep.db.Collection("todos").Find(ctx, bson.M{})
+func (rep *ToDoRepository) GetAll(ctx context.Context, userId string) ([]domain.Todo, error) {
+	userObjId, err := primitive.ObjectIDFromHex(userId)
 	if err != nil {
 		return nil, err
 	}
 
-	var todos []domain.Todo
+	cursor, err := rep.db.Collection("todos").Find(ctx, bson.M{"user_id": userObjId})
+	if err != nil {
+		return nil, err
+	}
+
+	todos := make([]domain.Todo, 0)
 	if err := cursor.All(ctx, &todos); err != nil {
 		return nil, err
 	}
@@ -51,12 +61,18 @@ func (rep *ToDoRepository) GetAll(ctx context.Context) ([]domain.Todo, error) {
 	return todos, nil
 }
 
-func (rep *ToDoRepository) Create(ctx context.Context, in domain.CreateTodoInput) (domain.Todo, error) {
+func (rep *ToDoRepository) Create(ctx context.Context, userId string, in domain.CreateTodoInput) (domain.Todo, error) {
+	userObjId, err := primitive.ObjectIDFromHex(userId)
+	if err != nil {
+		return domain.Todo{}, err
+	}
+
 	result, err := rep.db.Collection("todos").
 		InsertOne(ctx, bson.M{
 			"name":       in.Name,
 			"created_at": time.Now().Format(time.RFC3339),
 			"updated_at": time.Now().Format(time.RFC3339),
+			"user_id":    userObjId,
 		})
 
 	if err != nil {
@@ -77,14 +93,19 @@ func (rep *ToDoRepository) Create(ctx context.Context, in domain.CreateTodoInput
 	return todo, nil
 }
 
-func (rep *ToDoRepository) Update(ctx context.Context, id string, in domain.UpdateTodoInput) (domain.Todo, error) {
+func (rep *ToDoRepository) Update(ctx context.Context, id, userId string, in domain.UpdateTodoInput) (domain.Todo, error) {
 	objId, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
 		return domain.Todo{}, err
 	}
 
+	userObjId, err := primitive.ObjectIDFromHex(userId)
+	if err != nil {
+		return domain.Todo{}, err
+	}
+
 	update := bson.M{"$set": bson.M{"name": in.Name, "updated_at": time.Now().Format(time.RFC3339)}}
-	_, err = rep.db.Collection("todos").UpdateOne(ctx, bson.M{"_id": objId}, update)
+	_, err = rep.db.Collection("todos").UpdateOne(ctx, bson.M{"_id": objId, "user_id": userObjId}, update)
 
 	var todo domain.Todo
 	err = rep.db.Collection("todos").
@@ -96,16 +117,19 @@ func (rep *ToDoRepository) Update(ctx context.Context, id string, in domain.Upda
 	}
 
 	return todo, nil
-
-	return todo, nil
 }
 
-func (rep *ToDoRepository) Delete(ctx context.Context, id string) error {
+func (rep *ToDoRepository) Delete(ctx context.Context, id, userId string) error {
 	objId, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
 		return err
 	}
 
-	_, err = rep.db.Collection("todos").DeleteOne(ctx, bson.M{"_id": objId})
+	userObjId, err := primitive.ObjectIDFromHex(userId)
+	if err != nil {
+		return err
+	}
+
+	_, err = rep.db.Collection("todos").DeleteOne(ctx, bson.M{"_id": objId, "user_id": userObjId})
 	return err
 }
